+++ effects-cats/shared/src/main/scala/busymachines/pureharm/effects/package.scala
@@ -30,6 +30,5 @@ package object effects
     * This object is meant to bring in everything that is in cats + some extra, without burdening the user with two
     * different imports.
     */
-  @scala.annotation.nowarn
-  object implicits extends PureharmEffectsImplicits
+  object implicits extends PureharmEffectsSyntax
 }